Don't call back into the workspace host to force change contexts
This behavior causes all sorts of problems: removing a file (which
can now happen on a background thread) then tries to call into
all sorts of UI-affinitized APIs which are unavailable there.
Practically,  there's no reason we must do this anyways: if you do
remove a file in one context, we'll update our internal structures to
point to a new context. There's no real harm in that being temporarily
out of sync: either we'll pick the right context at random and the
user won't notice, or if we pick the wrong one, they can select the one
they want and everything will be updated again.

Fixes dotnet/roslyn#31019.