When responding to an IVsHierarchy change, only refresh needed files
This is an optimization that if you have 20 or 30 files open, and only
one of them came from a shared asset project, and that project is
changing context, we don't need to refresh all of the rest. Since the
refreshing is on the UI thread, we want that to be quick.