Cleanup the fetching of IVsFileChangeEx for the FileChangeWatcherProvider
The FileChangeWatcherProvider has to do some magic so we can hand
out a FileChangeWatcher that fetches the IVsFileChangeEx from the shell,
but not in a way that implicitly risks any marshal back to the UI thread
in a way that might deadlock. The odd bit was we were doing some of
that initialization in the VisualStudioWorkspace constructor itself
which didn't really make sense.

While I'm here, we can also remove the hack that was working around
a shell bug where we might deadlock if we called into the async service
provider from a background thread.