Add check to AbstractEditorFactory to load new WinForms designer if necessary
When the AbstractEditorFactory needs to load the WinForms designer, it must first check the
project's TFM. If the TFM is netcoreapp and is greater or equal to 3.0, it specifies a
different string to the IVSMDDesignerService.CreateDesignerLoader(...) to indicate that the
new WinForms designer should be instantiated.

Eventually, this code will move into the WinForms designer itself.